Small update

+++ modules/speed_test.py
@@ -238,13 +238,14 @@ async def speed_test_key(key, key_id, server_name):
             ss_local.terminate()
             logger.debug('ss_local.terminate (finally)')
 
-            try:
-                # В конце удаляем временный файл
-                os.remove(JSON_FILE)
-                os.remove(FILE)
+            if MODE != 3:
+                try:
+                    # В конце удаляем временный файл
+                    os.remove(JSON_FILE)
+                    os.remove(FILE)
 
-            except Exception as e:
-                logger.error(f'Ошибка удаления файла: {e}')
+                except Exception as e:
+                    logger.error(f'Ошибка удаления файла: {e}')
 
     else:   # Если ключа нет, то это localhost:
         try:
